+++ tests/test_doc_examples.py
@@ -73,9 +73,9 @@ class TestReadmeExamples(TestCase):
 
         corpus_scores, _ = evaluate(candidates, mult_references, metrics="dcase2023")
         # print(corpus_scores)
-        # dict containing the score of each metric: "meteor", "cider_d", "spice", "spider", "spider_fl", "fluerr"
+        # dict containing the score of each metric: "meteor", "cider_d", "spice", "spider", "spider_fl", "fer"
 
-        expected_keys = ["meteor", "cider_d", "spice", "spider", "spider_fl", "fluerr"]
+        expected_keys = ["meteor", "cider_d", "spice", "spider", "spider_fl", "fer"]
         self.assertTrue(set(corpus_scores.keys()).issuperset(expected_keys))
 
     def test_example_3(self) -> None: